プレイ中BPMの管理を改善
